requirements.txt
absl-py==0.15.0
aiohttp==3.8.1
aiosignal==1.2.0
albumentations==1.0.3
alembic==1.7.5
async-timeout==4.0.1
attrs==21.4.0
blinker==1.4
Bottleneck==1.3.2
brotlipy==0.7.0
cachetools==4.2.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.4
click==8.0.3
cloudpickle==2.0.0
cryptography==3.4.8
cycler==0.11.0
cytoolz==0.11.0
dask==2021.10.0
databricks-cli==0.16.2
docker==5.0.3
efficientnet-pytorch==0.6.3
entrypoints==0.3
Flask==2.0.2
fonttools==4.25.0
frozenlist==1.2.0
fsspec==2022.1.0
gitdb==4.0.9
GitPython==3.1.26
google-auth==1.33.0
google-auth-oauthlib==0.4.1
greenlet==1.1.2
grpcio==1.42.0
gunicorn==20.1.0
idna==3.3
imagecodecs==2021.8.26
imageio==2.13.1
imgaug==0.4.0
importlib-metadata==4.8.2
itsdangerous==2.0.1
Jinja2==3.0.3
joblib==1.1.0
kiwisolver==1.3.1
locket==0.2.1
Mako==1.1.6
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.3
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
mlflow==1.23.1
multidict==5.2.0
munch==2.5.0
munkres==1.1.4
networkx==2.6.3
numexpr==2.8.1
numpy==1.21.2
oauthlib==3.1.1
olefile==0.46
packaging==21.3
pandas==1.3.4
partd==1.2.0
Pillow==8.4.0
pip==21.2.4
pretrainedmodels==0.7.4
prometheus-client==0.13.0
prometheus-flask-exporter==0.18.7
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.1.0
pyOpenSSL==21.0.0
pyparsing==3.0.4
PyQt5==5.12.3
PyQt5_sip==4.19.18
PyQtChart==5.12
PyQtWebEngine==5.12.1
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2021.3
PyWavelets==1.1.1
PyYAML==6.0
querystring-parser==1.2.4
requests==2.27.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-image==0.18.3
scikit-learn==1.0.1
scipy==1.7.3
segmentation-models-pytorch==0.2.1
setuptools==58.0.4
Shapely==1.7.1
SimpleITK==2.1.1
six==1.16.0
smmap==5.0.0
SQLAlchemy==1.4.31
sqlparse==0.4.2
tabulate==0.8.9
tensorboard==2.6.0
tensorboard-data-server==0.6.0
tensorboard-plugin-wit==1.6.0
tensorboardX==2.2
threadpoolctl==2.2.0
tifffile==2021.7.2
timm==0.4.12
toolz==0.11.2
torch==1.10.1
torchaudio==0.10.1
torchvision==0.11.2
tornado==6.1
tqdm==4.62.3
typing-extensions==3.10.0.2
urllib3==1.26.7
websocket-client==1.2.3
Werkzeug==2.0.2
wheel==0.37.1
yarl==1.6.3
zipp==3.7.0
